backend/pyproject.toml
[tool.poetry]
name = "grafana-mcp-backend"
version = "1.0.0"
description = "FastAPI LLM Gateway + MCP Proxy for Grafana MCP Platform"
authors = ["Platform Engineering Team"]
readme = "README.md"
python = "^3.12"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {version = "^0.30.0", extras = ["standard"]}
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
httpx = "^0.27.0"
structlog = "^24.2.0"
sse-starlette = "^2.1.0"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
slowapi = "^0.1.9"
prometheus-fastapi-instrumentator = "^7.0.0"
# LLM providers
openai = "^1.35.0"
anthropic = "^0.28.0"
google-generativeai = "^0.7.0"
# MCP SDK
mcp = "^1.0.0"
# Cloud secret providers
boto3 = "^1.34.0"
azure-keyvault-secrets = "^4.8.0"
azure-identity = "^1.17.0"
google-cloud-secret-manager = "^2.20.0"
# Caching and utilities
cachetools = "^5.3.0"
tenacity = "^8.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
ruff = "^0.4.0"
mypy = "^1.10.0"
safety = "^3.2.0"
pip-audit = "^2.7.0"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=xml --cov-report=term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
